[upd]: add: 1) guest can add product to basket; 2) guest can get basket

+++ backend/internal/service/basket/delete_item.go
@@ -1,15 +1,37 @@
 package basket
 
-import "context"
+import (
+	"context"
 
-func (s *service) DeleteItem(ctx context.Context, itemID string, userID string) error {
-	basket, err := s.GetByUserID(ctx, userID)
-	if err != nil {
-		return err
-	}
+	"github.com/Fi44er/sdmedik/backend/internal/model"
+	"github.com/gofiber/fiber/v2/middleware/session"
+)
 
-	if err := s.basketItemRepo.Delete(ctx, itemID, basket.ID); err != nil {
-		return err
+func (s *service) DeleteItem(ctx context.Context, itemID string, userID string, sess *session.Session) error {
+	if userID == "" {
+		basket, err := s.GetByUserID(ctx, userID, sess)
+		if err != nil {
+			return err
+		}
+
+		if err := s.basketItemRepo.Delete(ctx, itemID, basket.ID); err != nil {
+			return err
+		}
+	} else {
+		basket := sess.Get("basket").(*model.Basket)
+		newBasket := []model.BasketItem{}
+		for _, item := range basket.Items {
+			if item.ID != itemID {
+				newBasket = append(newBasket, item)
+				break
+			}
+		}
+
+		basket.Items = newBasket
+		sess.Set("basket", basket)
+		if err := sess.Save(); err != nil {
+			return err
+		}
 	}
 
 	return nil
